refactor: move useSelectable to checkbox folder
